+++ packages/tswebextension/src/lib/mv2/background/index.ts
@@ -25,6 +25,7 @@ export * from './tabs';
 export * from './request';
 export * from '../../common';
 export * from './configuration';
+export { companiesDbService } from '../../common/companies-db-service';
 export { StealthActions } from '../../common/stealth-actions';
 export type { MessageHandlerMV2 } from './messages-api';
 export { MESSAGE_HANDLER_NAME } from '../../common/message-constants';