+++ packages/tswebextension/test/lib/mv2/background/request/request-blocking-api.test.ts
@@ -45,6 +45,7 @@ const getGetBlockingResponseParamsData = (
         rule: result.getBasicResult(),
         popupRule: result.getPopupRule(),
         referrerUrl: '',
+        requestId: '1',
         requestUrl,
         requestType,
         contentType,